requirements.txt
apex==0.9.10dev
av==15.0.0
datasets==3.2.0
decord==0.6.0
deepspeed==0.15.4
einops==0.8.1
eval==0.0.5
fastchat==0.1.0
flash_attn==2.7.2.post1
imageio==2.36.1
liger_kernel==0.6.1
lmdeploy==0.9.2
matplotlib==3.10.5
numpy==2.3.2
opencv_python==4.10.0.84
packaging==25.0
peft==0.10.0
Pillow==11.3.0
pycocotools==2.0.8
sentencepiece==0.1.99
termcolor==3.1.0
timm==0.9.12
tokenizers==0.21.4
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.37.2
